cleanup list entities tests
